+++ src/app/system/flags/hooks/useExperimentFlag.ts
@@ -1,5 +1,5 @@
 import { useFlag } from "@unleash/proxy-client-react"
-import { EXPERIMENT_NAME } from "app/utils/experiments/experiments"
+import { EXPERIMENT_NAME } from "app/system/flags/experiments"
 
 export function useExperimentFlag(name: EXPERIMENT_NAME) {
   return useFlag(name)
