+++ src/app/system/flags/hooks/useUnleashInitializer.ts
@@ -1,13 +1,10 @@
 import { useUnleashClient } from "@unleash/proxy-client-react"
 import { GlobalStore } from "app/store/GlobalStore"
-import { useUnleashListener } from "app/utils/useUnleashListener"
 import { useEffect } from "react"
 
-export const UnleashInitializer: React.FC = ({ children }) => {
+export const useUnleashInitializer = () => {
   const client = useUnleashClient()
   const userID = GlobalStore.useAppState((state) => state.auth.userID)
-
-  useUnleashListener()
 
   useEffect(() => {
     if (userID) {
@@ -24,6 +21,4 @@ export const UnleashInitializer: React.FC = ({ children }) => {
       client.stop()
     }
   }, [userID])
-
-  return <>{children}</>
 }
